+++ client/src/Engine/Renderer.js
@@ -1,4 +1,4 @@
-import { Simple1DNoise } from '../../../shared/utils/SimplexNoise';
+import { noise, noiseSeed } from '../../../shared/utils/noise';
 
 export default class Renderer {
 	constructor(layers, anchor) {
@@ -25,21 +25,18 @@ export default class Renderer {
 
 	scatterNode({ layerName, Class, options, number, seed }) {
 		const layer = this.layers.filter((l) => l.name === layerName)[0];
-		const noise = new Simple1DNoise(seed);
 
 		for (let i = 0; i < number; i++) {
 			const node = new Class(options);
 			this.nodes[`${layer}-${node.name}-n${i}`] = node;
 
-			const s = Math.abs(
-				noise.getVal(i * 100 + seed) * (node.scaleMultiplier || 1)
-			);
+			const s = Math.abs(noise(i * 100 + seed) * (node.scaleMultiplier || 1));
 
-			const o = noise.getVal(i * 100 + seed) + 0.2;
+			const o = noise(i * 100 + seed) + 0.2;
 
 			const p = {
-				x: Math.abs(noise.getVal(i * 10000 + seed)) * window.innerWidth * 0.7,
-				y: Math.abs(noise.getVal(i * 1000 + seed)) * 300,
+				x: Math.abs(noise(i * 10000 + seed)) * window.innerWidth * 0.7,
+				y: Math.abs(noise(i * 1000 + seed)) * 300,
 			};
 
 			node.HTML.style.transform = `scale(${s},${s}) translate(${p.x}px,${p.y}px) rotate(0rad)`;
