+++ server/src/Game/Collision.js
@@ -1,13 +1,16 @@
 import { DEFAULTS, REQUEST } from '../../../shared/Consts';
-import { makeOctaves, Simple1DNoise } from '../../../shared/utils/SimplexNoise';
+import {
+	noise as noisefunc,
+	noiseSeed,
+	makeOctaves,
+} from '../../../shared/utils/noise';
 
 export default class Collision {
 	constructor(seed, window) {
 		this.seed = seed;
 		this.window = window;
 		this.players = {};
-
-		this.noise = new Simple1DNoise(seed);
+		noiseSeed(seed);
 	}
 
 	setPlayers(players) {
@@ -16,7 +19,7 @@ export default class Collision {
 
 	didColide(player) {
 		const noise =
-			makeOctaves(this.noise.getVal, player.position.x, {
+			makeOctaves(noisefunc, player.position.x + this.window.w / 2, {
 				octaves: DEFAULTS.GENERATION.OCTAVES,
 				frequency: DEFAULTS.GENERATION.SCALE,
 				lacunarity: DEFAULTS.GENERATION.LACUNARITY,
@@ -33,6 +36,6 @@ export default class Collision {
 			);
 		}
 
-		return player.position.y + 25 / 2 >= noise;
+		return player.position.y + 60 >= noise;
 	}
 }
